fix: refactor gc entry aggregation to centralize domain rules
