+++ gradle-plugin/src/main/kotlin/io/github/cdsap/gcreport/plugin/output/DevelocityValues.kt
@@ -4,8 +4,9 @@ import com.gradle.develocity.agent.gradle.DevelocityConfiguration
 import io.github.cdsap.gcreport.plugin.GCReportExtension
 import io.github.cdsap.gcreport.plugin.extensions.getFileName
 import io.github.cdsap.gcreport.plugin.histogram.Histogram
-import io.github.cdsap.gcreport.plugin.model.GCCollectionMetrics
 import io.github.cdsap.gcreport.plugin.model.GCEntry
+import io.github.cdsap.gcreport.plugin.model.collectionTypeCounts
+import io.github.cdsap.gcreport.plugin.model.entriesForHistogram
 
 class DevelocityValues(
     private val develocityConfiguration: DevelocityConfiguration,
@@ -14,19 +15,18 @@ class DevelocityValues(
     private val extension: GCReportExtension,
 ) {
     fun report() {
-        val collectionMetrics = GCCollectionMetrics(gcEntries)
         develocityConfiguration.buildScan {
-            collectionMetrics.countsByDescription().forEach { (description, count) ->
+            gcEntries.collectionTypeCounts().forEach { (description, count) ->
                 value("gc-${log.getFileName()}-$description", "$count")
             }
-            val counter = collectionMetrics.totalCollections()
+            val counter = gcEntries.entriesForHistogram().size
             if (counter != 0) {
                 value("gc-${log.getFileName()}-total-collections", "$counter")
             }
             if (extension.histogramEnabled.get()) {
                 val histogram =
                     Histogram(extension.histogramBucket.get()).getHistogram(
-                        collectionMetrics.entriesForHistogram(),
+                        gcEntries.entriesForHistogram(),
                     )
                 var histogramText = "["
                 histogram.forEach {
